requirements.txt
fastapi==0.123.4
uvicorn==0.38.0

httpx==0.28.1
requests==2.32.5

streamlit==1.51.0

pysisense==0.2.0
mcp==1.24.0

python-dotenv==1.2.1
PyJWT==2.10.1
python-multipart==0.0.20

pydantic==2.12.4
pydantic-settings==2.12.0

sse-starlette==3.0.3

pandas==2.3.3
numpy==2.3.4
PyYAML==6.0.3

anyio==4.11.0